issue-61 - simplify columns block by using cards block
